Make data/mod.rs contains only reexports
